es_vec_classifier/requirements.txt
requests>=2.22.0
tensorflow>=2.1.0
elasticsearch==7.5.0
numpy>=1.19.3
tensorflow_hub>=0.8.0
sentence_transformers>=0.4.1.2
progressbar33>=2.4
tensorflow_text>=2.4.3
